# openapi.yaml
openapi: 3.1.0
info:
  title: NeroCD API
  version: 0.1.0
  description: |
    Public REST API for the NeroCD automation control plane.

    All public routes are versioned under `/api/v1`. List responses use a stable
    `{ "items": [...] }` envelope. Error responses use
    `{ "code": "stable_code", "error": "message" }`.

    List endpoints return a stable `{ "items": [...] }` envelope plus
    pagination metadata: `limit`, `offset`, `count`, and `total`. Clients can
    request `limit` and `offset` query parameters. Filtering is documented per
    endpoint. Mutations consume JSON request bodies and return the created
    resource or validation errors.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
security:
  - bearerAuth: []
  - browserSession: []
paths:
...
  /api/v1/runners/deployments/provenance:
    post:
      summary: Bind runner-observed immutable deployment provenance
      operationId: resolveRunnerDeploymentProvenance
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ProvenanceResolutionInput"}}}}
      responses:
        "200": {description: Provenance acknowledged, content: {application/json: {schema: {$ref: "#/components/schemas/Revision"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
...
  /api/v1/runners/events/batch:
    post:
      summary: Append an idempotent bounded batch of attempt events
      operationId: appendRunnerEvents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunEventBatchInput"
      responses:
        "200":
          description: Authoritative persisted events, in request order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEventBatchAck"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/v1/runners/secrets/access:
    post:
      summary: Authorize and audit one fenced runner-local secret access
      operationId: authorizeRunnerSecretAccess
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretAccessInput"
      responses:
        "200":
          description: Safe idempotent secret-access authorization metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretAccessGrant"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/v1/runners:
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      summary: List runners
      operationId: listRunners
      responses:
        "200":
          description: Runner list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerListResponse"
        "401":
          $ref: "#/components/responses/Error"
...
  /api/v1/runner-enrollments/consume:
    post:
      summary: Consume an enrollment using its one-time bearer token
      operationId: consumeRunnerEnrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerEnrollmentConsumeInput"
      responses:
        "200":
          description: Bound runner identity. The long-lived credential is generated and retained only by the runner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumedRunnerEnrollment"
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/runners/renew:
    post:
      summary: Renew a fenced runner lease
      operationId: renewRunnerLease
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lease_id, attempt, fence]
              properties:
                lease_id: {type: string}
                attempt: {type: integer, minimum: 1}
                fence: {type: string}
      responses:
        "200":
          description: Lease renewed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLease"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/runners/complete:
    post:
      summary: Complete run lease
      operationId: completeLease
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lease_id, attempt, fence, completion_key, status]
              properties:
                lease_id:
                  type: string
                attempt:
                  type: integer
                  minimum: 1
                fence:
                  type: string
                completion_key:
                  type: string
                status:
                  type: string
                  enum: [succeeded, failed, canceled]
      responses:
        "200":
          description: Completed lease.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunLease"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: NeroCD session token
      description: CLI session and API token credential.
    browserSession:
      type: apiKey
      in: cookie
      name: nerocd_session
      description: "Browser-only local session cookie. Unsafe cookie-authenticated requests require X-NeroCD-CSRF: 1."
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return (0-100). Omit to use the bounded default.
      schema:
        type: integer
        minimum: 0
        maximum: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip from the filtered result.
      schema:
        type: integer
        minimum: 0
        maximum: 100000
  responses:
    Error:
      description: Stable error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Revision:
      type: object
      required: [id, service_id, requested_ref, git_commit, compose_hash, image_digests, content_identity, provenance_state, provenance_resolved, created_by, created_at]
      properties:
        id: {type: string}
        service_id: {type: string}
        requested_ref: {type: string}
        git_commit: {type: string}
        compose_hash: {type: string}
        image_digests: {type: array, items: {type: string}}
        content_identity: {type: string}
        provenance_state: {type: string, enum: [pending, resolved, legacy_unverified]}
        provenance_resolved: {type: boolean}
        resolved_at: {type: [string, "null"], format: date-time}
        created_by: {type: string}
        created_at: {type: string, format: date-time}
...
    ProvenanceResolutionInput:
      type: object
      required: [deployment_id, run_id, lease_id, attempt, fence, resolution_id, git_commit, compose_hash, image_digests, content_identity]
      properties:
        deployment_id: {type: string}
        run_id: {type: string}
        lease_id: {type: string}
        attempt: {type: integer, minimum: 1}
        fence: {type: string, minLength: 1}
        resolution_id: {type: string, minLength: 1}
        git_commit: {type: string, pattern: '^[0-9a-f]{40,64}$'}
        compose_hash: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        image_digests: {type: array, minItems: 1, items: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}}
        content_identity: {type: string}
...
    ErrorResponse:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          enum: [unauthenticated, forbidden, not_found, bad_request, internal_error]
        error:
          type: string
...
    SecretAccessInput:
      type: object
      required: [access_id, run_id, lease_id, attempt, fence, binding, provider]
      properties:
        access_id:
          type: string
          pattern: '^secret_access_[0-9a-f]{32}$'
        run_id:
          type: string
        lease_id:
          type: string
        attempt:
          type: integer
          minimum: 1
        fence:
          type: string
        binding:
          type: string
        provider:
          type: string
          enum: [runner_file, env]
        version:
          type: string
    SecretAccessGrant:
      type: object
      required: [access_id, run_id, lease_id, attempt, binding, provider, authorized_at]
      properties:
        access_id:
          type: string
        run_id:
          type: string
        lease_id:
          type: string
        attempt:
          type: integer
        binding:
          type: string
        provider:
          type: string
        version:
          type: string
        authorized_at:
          type: string
          format: date-time
...
    Runner:
      type: object
      required: [id, name, tags, capabilities, status, registered_at, last_heartbeat_at]
      properties:
        id:
          type: string
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        status:
          type: string
        registered_at:
          type: string
          format: date-time
        last_heartbeat_at:
          type: string
          format: date-time
...
    RunnerEnrollmentConsumeInput:
      type: object
      required: [request_id, credential_hash]
      properties:
        request_id: { type: string, pattern: '^enroll_consume_[0-9a-f]{32}$' }
        credential_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
...
    ConsumedRunnerEnrollment:
      type: object
      required: [runner]
      properties:
        runner: { $ref: "#/components/schemas/Runner" }
...
    RunnerListResponse:
      type: object
      required: [items, limit, offset, count, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Runner"
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        count:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    RunLease:
      type: object
      required: [id, run_id, runner_id, status, expires_at, created_at, attempt, fence]
      properties:
        id:
          type: string
        run_id:
          type: string
        runner_id:
          type: string
        status:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        attempt:
          type: integer
          minimum: 1
        fence:
          type: string
...
    RunLog:
      type: object
      required: [id, run_id, sequence, stream, message, created_at]
      properties:
        id:
          type: string
        run_id:
          type: string
        sequence:
          type: integer
        stream:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
        event_key:
          type: string
        lease_id:
          type: string
        attempt:
          type: integer
          minimum: 1
        requested_sequence:
          type: integer
          minimum: 1
...
    RunEventInput:
      type: object
      required: [event_key, sequence, stream, message]
      properties:
        event_key: {type: string}
        sequence: {type: integer, minimum: 1}
        stream: {type: string, enum: [system, stdout, stderr]}
        message: {type: string}
    RunEventBatchInput:
      type: object
      required: [run_id, lease_id, attempt, fence, events]
      properties:
        run_id: {type: string}
        lease_id: {type: string}
        attempt: {type: integer, minimum: 1}
        fence: {type: string}
        events:
          type: array
          minItems: 1
          maxItems: 64
          items:
            $ref: "#/components/schemas/RunEventInput"
    RunEventBatchAck:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RunLog"
...